Feat: 로드시 태그 상태 적용
로드했을 때 태그 상태가 적용되도록 기능 추가
